Cleanup - move types to its own folder
