+++ src/components/PomoForm.tsx
@@ -15,7 +15,7 @@ import {
   POMO_FORM_NAME_CHECKBOX,
   POMO_FORM_NAME_TEXTFIELD,
 } from '../constants';
-import { PomoFormProps } from '../constants/types';
+import { PomoFormProps } from '../types/';
 import './PomoForm.less';
 
 export default function PomoForm({ onSubmitForm }: PomoFormProps) {
