+++ src/components/TodoForm.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { Checkbox, TextField } from '@mui/material';
-import { TodoFormProps } from '../constants/types';
+import { TodoFormProps } from '../types/';
 import { TODO_FORM_HELPER_TEXT_FIELD, TODO_FORM_NAME_CHECKBOX, TODO_FORM_NAME_TEXTFIELD } from '../constants';
 import './TodoForm.less';
